+++ lib/libvmmapi/amd64/Makefile.inc
@@ -1,2 +1,3 @@
-SRCS+=	vmmapi_machdep.c	\
+SRCS+=	ppt.c			\
+	vmmapi_machdep.c	\
 	vmmapi_freebsd_machdep.c
