+++ lib/libvmmapi/vmmapi.c
@@ -742,113 +742,6 @@ vm_set_capability(struct vcpu *vcpu, enum vm_cap_type cap, int val)
 	return (vcpu_ioctl(vcpu, VM_SET_CAPABILITY, &vmcap));
 }
 
-int
-vm_assign_pptdev(struct vmctx *ctx, int bus, int slot, int func)
-{
-	struct vm_pptdev pptdev;
-
-	bzero(&pptdev, sizeof(pptdev));
-	pptdev.bus = bus;
-	pptdev.slot = slot;
-	pptdev.func = func;
-
-	return (ioctl(ctx->fd, VM_BIND_PPTDEV, &pptdev));
-}
-
-int
-vm_unassign_pptdev(struct vmctx *ctx, int bus, int slot, int func)
-{
-	struct vm_pptdev pptdev;
-
-	bzero(&pptdev, sizeof(pptdev));
-	pptdev.bus = bus;
-	pptdev.slot = slot;
-	pptdev.func = func;
-
-	return (ioctl(ctx->fd, VM_UNBIND_PPTDEV, &pptdev));
-}
-
-int
-vm_map_pptdev_mmio(struct vmctx *ctx, int bus, int slot, int func,
-		   vm_paddr_t gpa, size_t len, vm_paddr_t hpa)
-{
-	struct vm_pptdev_mmio pptmmio;
-
-	bzero(&pptmmio, sizeof(pptmmio));
-	pptmmio.bus = bus;
-	pptmmio.slot = slot;
-	pptmmio.func = func;
-	pptmmio.gpa = gpa;
-	pptmmio.len = len;
-	pptmmio.hpa = hpa;
-
-	return (ioctl(ctx->fd, VM_MAP_PPTDEV_MMIO, &pptmmio));
-}
-
-int
-vm_unmap_pptdev_mmio(struct vmctx *ctx, int bus, int slot, int func,
-		     vm_paddr_t gpa, size_t len)
-{
-	struct vm_pptdev_mmio pptmmio;
-
-	bzero(&pptmmio, sizeof(pptmmio));
-	pptmmio.bus = bus;
-	pptmmio.slot = slot;
-	pptmmio.func = func;
-	pptmmio.gpa = gpa;
-	pptmmio.len = len;
-
-	return (ioctl(ctx->fd, VM_UNMAP_PPTDEV_MMIO, &pptmmio));
-}
-
-int
-vm_setup_pptdev_msi(struct vmctx *ctx, int bus, int slot, int func,
-    uint64_t addr, uint64_t msg, int numvec)
-{
-	struct vm_pptdev_msi pptmsi;
-
-	bzero(&pptmsi, sizeof(pptmsi));
-	pptmsi.bus = bus;
-	pptmsi.slot = slot;
-	pptmsi.func = func;
-	pptmsi.msg = msg;
-	pptmsi.addr = addr;
-	pptmsi.numvec = numvec;
-
-	return (ioctl(ctx->fd, VM_PPTDEV_MSI, &pptmsi));
-}
-
-int	
-vm_setup_pptdev_msix(struct vmctx *ctx, int bus, int slot, int func,
-    int idx, uint64_t addr, uint64_t msg, uint32_t vector_control)
-{
-	struct vm_pptdev_msix pptmsix;
-
-	bzero(&pptmsix, sizeof(pptmsix));
-	pptmsix.bus = bus;
-	pptmsix.slot = slot;
-	pptmsix.func = func;
-	pptmsix.idx = idx;
-	pptmsix.msg = msg;
-	pptmsix.addr = addr;
-	pptmsix.vector_control = vector_control;
-
-	return ioctl(ctx->fd, VM_PPTDEV_MSIX, &pptmsix);
-}
-
-int
-vm_disable_pptdev_msix(struct vmctx *ctx, int bus, int slot, int func)
-{
-	struct vm_pptdev ppt;
-
-	bzero(&ppt, sizeof(ppt));
-	ppt.bus = bus;
-	ppt.slot = slot;
-	ppt.func = func;
-
-	return ioctl(ctx->fd, VM_PPTDEV_DISABLE_MSIX, &ppt);
-}
-
 uint64_t *
 vm_get_stats(struct vcpu *vcpu, struct timeval *ret_tv,
 	     int *ret_entries)
